Move content of `grpcutil/` one level up
