+++ middlewares/metrics/prometheus.go
@@ -4,8 +4,8 @@ import (
 	grpc_prometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
 	"google.golang.org/grpc"
 
-	"github.com/aserto-dev/aserto-grpc/grpcutil"
-	config "github.com/aserto-dev/aserto-grpc/grpcutil/metrics"
+	grpcutil "github.com/aserto-dev/aserto-grpc"
+	config "github.com/aserto-dev/aserto-grpc/metrics"
 )
 
 func NewMiddlewares(conf config.Config, middlewares ...grpcutil.Middleware) grpcutil.Middlewares {
